# operator-api/openapi.yaml
openapi: 3.1.0
info:
  title: Kernel Images API
  version: 0.1.0
paths:
...
  /process/exec:
    post:
      summary: Execute a command synchronously (optional streaming)
      operationId: processExec
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcessExecRequest" }
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessExecResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /process/spawn:
    post:
      summary: Execute a command asynchronously
      operationId: processSpawn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcessSpawnRequest" }
      responses:
        "200":
          description: Spawned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessSpawnResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /process/{process_id}/status:
    get:
      summary: Get process status
      operationId: processStatus
      parameters:
        - in: path
          name: process_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessStatusResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /process/{process_id}/stdout/stream:
    get:
      summary: Stream process stdout/stderr (SSE)
      operationId: processStdoutStream
      parameters:
        - in: path
          name: process_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: text/event-stream of stdout/stderr
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"stream":"stdout","data_b64":"SGVsbG8K"}
                  event: data
                  data: {"event":"exit","exit_code":0}
        "404":
          $ref: "#/components/responses/NotFoundError"
  /process/{process_id}/stdin:
    post:
      summary: Write to process stdin
      operationId: processStdin
      parameters:
        - in: path
          name: process_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcessStdinRequest" }
      responses:
        "200":
          description: Bytes written
          content:
            application/json:
              schema:
                type: object
                properties:
                  written_bytes: { type: integer }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /process/{process_id}/kill:
    post:
      summary: Send signal to process
      operationId: processKill
      parameters:
        - in: path
          name: process_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcessKillRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /network/proxy/socks5/start:
    post:
      summary: Start a lightweight SOCKS5 proxy in the VM
      operationId: networkSocksStart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Socks5StartRequest" }
      responses:
        "200":
          description: Proxy started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Socks5StartResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /network/proxy/socks5/stop:
    post:
      summary: Stop a SOCKS5 proxy
      operationId: networkSocksStop
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Socks5StopRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /network/intercept/rules:
    post:
      summary: Apply realtime network intercept rules
      operationId: networkInterceptApply
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NetworkInterceptRulesRequest" }
      responses:
        "200":
          description: Rule set applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NetworkInterceptRulesResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /network/intercept/rules/{rule_set_id}:
    delete:
      summary: Remove intercept rule set
      operationId: networkInterceptDelete
      parameters:
        - in: path
          name: rule_set_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /network/forward:
    post:
      summary: Configure port forwarding host↔VM
      operationId: networkForward
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForwardRequest" }
      responses:
        "200":
          description: Forward active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForwardResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /network/forward/{forward_id}:
    delete:
      summary: Remove a port forward
      operationId: networkForwardDelete
      parameters:
        - in: path
          name: forward_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /logs/stream:
    get:
      summary: Subscribe to logs via SSE
      operationId: logsStream
      parameters:
        - in: query
          name: source
          required: true
          schema:
            type: string
            enum: [kernel, syslog, application, path]
        - in: query
          name: path
          required: false
          schema: { type: string }
        - in: query
          name: follow
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: text/event-stream of log lines
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"source":"path","line":"started","ts":"2025-08-09T10:00:00Z"}
...
  /macros/create:
    post:
      summary: Create a macro from steps
      operationId: macrosCreate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MacroCreateRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MacroCreateResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /macros/run:
    post:
      summary: Run a macro
      operationId: macrosRun
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MacroRunRequest" }
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MacroRunResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /macros/{macro_id}:
    delete:
      summary: Delete a macro
      operationId: macrosDelete
      parameters:
        - in: path
          name: macro_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /os/locale:
    get:
      summary: Get OS locale, keyboard layout, timezone
      operationId: osLocaleGet
      responses:
        "200":
          description: Locale info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocaleInfo" }
    post:
      summary: Set OS locale, keyboard layout, timezone
      operationId: osLocaleSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocaleUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocaleUpdateResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
...
  /pipe/send:
    post:
      summary: Send a JSON object onto a named pipe channel
      operationId: pipeSend
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PipeSendRequest" }
      responses:
        "200":
          description: Enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipeSendResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /pipe/recv/stream:
    get:
      summary: Receive JSON objects from a named pipe channel (SSE)
      operationId: pipeRecvStream
      parameters:
        - in: query
          name: channel
          required: true
          schema: { type: string, default: "default" }
      responses:
        "200":
          description: text/event-stream of objects
          headers:
            X-SSE-Content-Type:
              description: Media type of SSE data events (application/json)
              schema:
                type: string
                const: application/json
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: data
                  data: {"ts":"2025-08-09T10:00:00Z","object":{"event":"ready"}}
...
components:
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
...
    ProcessExecRequest:
      type: object
      properties:
        command: { type: string }
        args: { type: array, items: { type: string }, default: [] }
        cwd: { type: string, nullable: true }
        env: { type: object, additionalProperties: { type: string }, default: {} }
        as_user: { type: string, nullable: true }
        as_root: { type: boolean, default: false }
        timeout_sec: { type: integer, nullable: true }
        stream: { type: boolean, default: false }
      required: [command]
    ProcessExecResponse:
      type: object
      properties:
        exit_code: { type: integer }
        stdout_b64: { type: string }
        stderr_b64: { type: string }
        duration_ms: { type: integer }
    ProcessSpawnRequest:
      allOf:
        - $ref: "#/components/schemas/ProcessExecRequest"
      properties:
        stream: { readOnly: true }
    ProcessSpawnResponse:
      type: object
      properties:
        process_id: { type: string, format: uuid }
        pid: { type: integer }
        started_at: { type: string, format: date-time }
    ProcessStatusResponse:
      type: object
      properties:
        state: { type: string, enum: [running, exited] }
        exit_code: { type: integer, nullable: true }
        cpu_pct: { type: number }
        mem_bytes: { type: integer }
    ProcessStdinRequest:
      type: object
      properties:
        data_b64: { type: string }
      required: [data_b64]
    ProcessKillRequest:
      type: object
      properties:
        signal: { type: string, enum: [TERM, KILL, INT, HUP], default: TERM }
      required: [signal]
    Socks5StartRequest:
      type: object
      properties:
        bind_host: { type: string, default: "127.0.0.1" }
        bind_port: { type: integer, default: 1080 }
        auth:
          type: object
          nullable: true
          properties:
            username: { type: string, nullable: true }
            password: { type: string, nullable: true }
      required: [bind_host, bind_port]
    Socks5StartResponse:
      type: object
      properties:
        proxy_id: { type: string, format: uuid }
        url: { type: string, example: "socks5://127.0.0.1:1080" }
    Socks5StopRequest:
      type: object
      properties:
        proxy_id: { type: string, format: uuid }
      required: [proxy_id]
    NetworkInterceptRulesRequest:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/NetworkRule"
      required: [rules]
    NetworkInterceptRulesResponse:
      type: object
      properties:
        rule_set_id: { type: string, format: uuid }
        applied: { type: boolean }
    NetworkRule:
      type: object
      properties:
        id: { type: string }
        match:
          type: object
          properties:
            protocol: { type: string, enum: [http, https], nullable: true }
            host_contains: { type: string, nullable: true }
            path_regex: { type: string, nullable: true }
            method: { type: string, enum: [GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS], nullable: true }
        action:
          type: object
          properties:
            type: { type: string, enum: [block, delay, modify_request, modify_response, mock_response] }
            delay_ms: { type: integer, nullable: true }
            set_request_headers: { type: object, additionalProperties: { type: string }, nullable: true }
            set_response_headers: { type: object, additionalProperties: { type: string }, nullable: true }
            status: { type: integer, nullable: true }
            body_b64: { type: string, nullable: true }
      required: [id, match, action]
...
    ForwardRequest:
      type: object
      properties:
        direction: { type: string, enum: [host_to_vm, vm_to_host] }
        host_port: { type: integer }
        vm_port: { type: integer }
      required: [direction, host_port, vm_port]
    ForwardResponse:
      type: object
      properties:
        forward_id: { type: string, format: uuid }
        active: { type: boolean }
...
    MacroCreateRequest:
      type: object
      properties:
        name: { type: string }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/MacroStep"
      required: [name, steps]
    MacroStep:
      type: object
      properties:
        action: { type: string, example: keyboard.type }
        text: { type: string, nullable: true }
        key: { type: string, nullable: true }
        ms: { type: integer, nullable: true, description: for sleep action }
      required: [action]
    MacroCreateResponse:
      type: object
      properties:
        macro_id: { type: string, format: uuid }
    MacroRunRequest:
      type: object
      properties:
        macro_id: { type: string, format: uuid }
      required: [macro_id]
    MacroRunResponse:
      type: object
      properties:
        started: { type: boolean }
        run_id: { type: string, format: uuid }
...
    LocaleInfo:
      type: object
      properties:
        locale: { type: string, example: en_US.UTF-8 }
        keyboard_layout: { type: string, example: us }
        timezone: { type: string, example: UTC }
    LocaleUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/LocaleInfo"
      properties:
        apply_to: { type: string, enum: [current_session, system], default: current_session }
    LocaleUpdateResponse:
      type: object
      properties:
        updated: { type: boolean }
        requires_restart: { type: boolean }
...
    PipeSendRequest:
      type: object
      properties:
        channel: { type: string, default: default }
        object: { type: object }
      required: [channel, object]
    PipeSendResponse:
      type: object
      properties:
        enqueued: { type: boolean }
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...